Refactor the code to be more extensible
This commit encompasses changes to the package so that it can handle
snippet expansions in a more easily extensible manner than the original
prototype.

To do this we pushed the parser for the snippets out into a separate
function and have it use a new list of possible extensions to know if
a snippet is of interest (and if so, how).