Add format support for custom variable expansions
This is a small change that allows for custom variables to take some
form of custom format to control how they expand.

To do this, the contract on the enhancement class has been modified so
that along with the list of variables, the (potentially) modified
content of the snippet is also returned.

For the case of our sample DATE expansion, this allows for full format
support by finding something like ${DATE:%Y-%m-%d} and outputting it as
a custom variable (e.g. ${DATE_1}) which has the appropriately
formatted value.