Split the single plugin into a proper code tree
This commit leaves the code unchanged functionally (but see below) BUT
does restructure things so that instead of one giant hard to read
plugin file, we instead have a more robust file tree.

This splits the code into a library of common routines, a source tree,
and within the source tree a folder for commands. Everything is set to
dynamically reload if the core plugin reloads to make sure that
everything is up to date.

Although the code is functionally unchanged, some small tweaks were made
to include the stubs of a new API, which was needed because various
things were tracked in global variables.