Alter the list of extensions to be instances
Previously the list of extensions was stored as a list of class objects
and we invoked class methods on them. This has been changed and cleaned
up so that we now store instances of the classes instead.

In addition, this alters the API contract on the base class so that it
does the heavy lifting for most cases by doing the variable search for
us. This greatly cleans up the existing examples.