Implement our initial SnippetManager class
This is an implementation of our initial SnippetManager class, which is
responsible for loading, scanning and searching the list of snippets
that we are enhancing.

In implementing this we also fixed the list of completion enhancement
classes not being dynamic enough to replace classes if they're
registered more than once, since AutomaticPackageReloader reloads files
in a way that added each item to the list many times.