Implement our initial CompletionManager class
This management class is for keeping track of the registered enhancement
classes that can be used to augment snippet variables.

There is more work to be done on this, since we need some way to allow
the snippet manager to know that it should rescan snippets (or we could
brute force that from within here but I don't like that) but this is a
good step towards making the prototype code cleaner.