Focus the ability to reload a snippet
Previously we had a simple event listener listening for saves of any
snippet file and causing an entire snippet cache refresh to bring that
change in.

Now that we have a snippet manager we can clean up a great deal by
having a more focused reload of just the snippet and nothing else; this
includes being able to detect when a new snippet is created, and when
the content of a snippet makes it be enhanced or not.