Watch for settings changes and apply snippets
This is a rather large change that changes a few of the internals to be
clearer in order to make the ultimate goal easier, which is to have a
settings listener that watches the ignored_packages setting.

With this change in place we now watch to see when packages are being
ignored and unignored, and either discard snippets from them or load
them as appropriate.

This required some internal changes because various things were being
created on module load, but since settings listeners are now at the
core and they require the plugin_loaded method to be called, things
needed to be rearranged.