Set up for our own custom enhanced snippet files
The previous iterations of this package piggybacked on the existing
snippet file format, and only tried to "enhance" them if they contained
special variables.

This is OK, but it requires that we write code that adds such snippets
to the list of ignored snippets for packages so that both versions don't
appear in the AC panel.

This change sets us up with a new resource type for enhanced snippets,
and we just load them all unconditionally.