Implement a first version of external enhancements
In this initial version of the code, the enhancements manager can scan
for all packages that contain a sentinel file that indicates that they
contain enhancements, and then dynamically load those files and grab
the enhancements out.

This is implemented such that we import (and possibly reload if it was
previously loaded) a given module and then grab from its module table
any classes that represent snippet enhancements.