Add the ability to trigger snippets by filename
This augments our enhanced snippets so that not only can you specify a
scope to say where they apply, you can also apply an optional glob
pattern that must match the current filename in order to expand out.

For example, you could have a snippet that only expands in files of a
certain name, only in a scope, and only in a scope if that file matches
a given name (e.g. you have a snippet of Python code that you only want
to be active while working on test cases).