Change how snippets and enhancements interact
Previously snippets would, at load time, capture a static list of the
classes that represent their variables for later expansion. A problem
with this is that when a snippet reloads, we would be forced to reload
all enhancements, and vice versa, because they are tightly coupled.

This change decouples that so that the list of enhancements is
searchable by field name, and snippets store the list of their fields,
which they look up at injection time.

As a result either list can be reloaded without requiring a load of the
other and everything still associates properly.