Include a YAML based snippet format
This commit vendors in the python-frontmatter package and also the
current pyyaml dependency available on package control, which is
currently not available in the 3.8 plugin host, which we require.

With this, we gain the ability to have enhanced snippets in a non-XML
format by treating them as text files with a YAML frontmatter.

The snippet loader has been enhanced to try to load first XML formatted
snippets and then the new format as a fallback, allowing us to handle
snippets of both types.