Include a YAML syntax for our snippets
This is entirely hacky and gross in almost every conceivable way, plus
it is not complete, but it shows the proof of concept well enough.

The syntax inherits from the YAML syntax and then reuses contexts to
make the YAML apply only to frontmatter, using custom highlighting for
the remainder of the file.

A small plugin change is also made to the event handler so that when a
file is loaded, anything that looks like YAML will get that syntax
applied. On save, the syntax will flip between the two syntaxes that are
available based on content.