Implement our own insert_enhanced_snippet command
This command is a drop in replacement for the internal insert_snippet
command (which this one actually uses behind the scenes), and can source
snippets from inline text, pre-loaded enhanced snippets, as well as
regular snippets if their names are given.

In all cases, any enhanced variables will be expanded out as a part of
the operation.