Include a setting use_details
When this setting is turned on, the completions that we insert will be
marked with a details key, which causes the AC panel to display an extra
row that shows the details on the kind of symbol this is.

Since not everyone likes this, this allows you to turn that off
(though if other things are going to use details, you might want to
turn it on).