New Snippet load and application logic
This commit comprises two changes:

When creating a snippet from just content, the snippet loader can now
also apply a scope and glob if one is provided, to allow for a text
string as well as providing contextual clues for where the snippet
belongs.

Also, there was a single piece of code in the snippet manager that knows
how to determine if a snippet applies; soon that is going to be
something that we need to do in code outside the manager too.

This logic was factored out into a shared method that can be accessed
elsewhere.