Include a new setting for debug logging
Since there are a lot of logs that are of no use to people that are
using the package, this adds in a new debug setting that controls
wether debug logs are printed or not, as well as a new debug logger
which respects the settings.

Various logs were cleaned up and many now use the debug logger so that
a setting can turn off a lot of the noise.