Rename snippet fields to variables
This was a bit of a misname in the original version of the code, but the
member in the snippet is holding the names of variables, not the names
of fields, which are different (and about to be added).

This change renames that field to be variables instead, so that the
upcoming change is less confusing.