Include numeric fields in Snippet info object
The upcoming changes for field options require that we know about the
total number of numeric fields that exist in the snippet, so that as
the options are navigated, we know where we are.

This change includes the simple ability to grab that information out of
the snippet body content.