Load and parse optional field value lists
In order to provide support for numbered fields having optional lists of
values to be inserted when they are focused, this set of changes allows
us a (currently crude) method of pulling such options, their field
assignments and their placeholders and values.

This is currently in a WBG state (working, but gross) and does not have
any validation currently. However this allows us to get it into the
hands of people to play with and find issues, and they may have
usability ideas that change how we handle this data entirely (or what
its structure is).