Implement multiple choice snippet field values
Bringing the lat two commits together, this modifies the completion
handling so that fields that have multiple choice options will prompt
the user for an option from the list.

This is still in early alpha and make have some edge cases; in
particular keep an eye out for things that are marked for being finessed
into something that is less garbage like.