Fix an issue with the XML snippet loader
During refactoring of the XML snippet loader as a part of the optional
field handling, an issue was added to the XML loader that went unnoticed
in which it would always fail to load the file because it was attempting
to fetch field values from a value that did not exist.

The result of this is that all XML snippets were treated as plain text
snippets with no tab triggers or scopes.