Include a command for generating snippets
This small change implements a command that will generate a new empty
enhanced snippet stub in the same manner as the built in snippet
creation does.

The inserted template has commented out versions of all of the possible
keys that can be presented, as well as a quick demo of optional field
values.