Augment how field options are specified in YAML
This changes how we look for and handle the options for fields in an
enhanced YAML snippet (this does not affect XML snippets, which do not
have this functionality currently as support for them may be removed in
an upcoming release).

Previously we would fetch a list of strings and use those for the values
of fields with multiple potential values. This has been augmented such
that the user can provide not only text, but also details to be
displayed within the quick panel to describe what each item is for.