Defer settings listener setup
This commit may ultimately not be needed, but it was observed that under
Linux, when the package loads from a sublime-package file, the call to
load_settings() returns a settings object with ID 0 that responds to all
method calls by returning None.

This quick hacky tweak defers the init until the current task in the
queue is completed, to give the file catalog a chance to catch up with
the catalog update.