Re-introduce legacy snippet loader
For upcoming changes, we want to be able to load snippets in the legacy
XML format, so that we can provide a conversion to the newer YAML based
format.

To that end, this re-introduces a form of the original XML loader, now
with the ability to load ad-hoc files from disk as well as resources and
static content.

We also are including here a custom frontmatter handler for handling our
snippets which is based on the default YAML based formatter, so that we
have more control over the load and generation of content going
forward.