Include a command to convert a snippet
This change introduces a new command which can convert an existing XML
snippet in the standard format into one of the new enhanced snippet
formatted files.

This will work for package resources that are both packed and unpacked,
files of the appropriate type that are not rooted in the packages
folder, and unsaved buffers that appear to be snippets.

The new command appears both in the main menu and in the context menu,
though in the context menu the command is disabled if it doesn't apply.