Include support for Docsify documentation
This modifies the structure of the repository to include a new docs/
folder within which lives the documentation for the project, which will
be served at runtime via Docsify.

The git attributes are adjusted to exclude the HTML and images from the
resulting package, since they are only needed for the documentation
site.