Fix the permalink to the animated demo gif
The README was accidentally left with a broken link to the animated
gif that shows a demo of the package in action. This is only an issue
because the README is used on packagecontrol.io on the package page,
and it has no relative file path to get at the image.

The link has been moved from a relative link to an absolute link to the
blob.